day 35 commit
Changed build script to be path agnostic
updated tm_print.h to v0.0.4a
changed how win32 layer acquires the dll path, it now also is path
agnostic
updated Rect.h to deduce common types of passed in arguments for the
return type. This way no error message is generated when trying to pass
int literals to functions that expect floats for instance.

+++ src/Core/StringView.cpp
@@ -253,8 +253,8 @@ StringViewPos findLast( StringView str, StringView findStr, int32 start = String
 	// check whether there is room for a match
 	if( start > 0 && str.size() >= findStr.size() ) {
 		size_t strSize = (size_t)str.size();
-		if( start < strSize ) {
-			strSize = start;
+		if( (size_t)start < strSize ) {
+			strSize = (size_t)start;
 		}
 		const char* current;
 		size_t findStrCount = (size_t)findStr.size();
